style updates; moved relay functionality to relay selectoin widget
