+++ src/styles/global.css
@@ -218,6 +218,7 @@ footer {
     flex-direction: column;
 }
 
+
 .single-card-profile-img {
     width: 150px;
     height: 150px;
@@ -231,7 +232,7 @@ footer {
     width: 70%;
     margin: 0 auto;
     text-align: justify;
-    margin-bottom: 4.5em;
+    margin-bottom: 90px;
     font-size: 1.2em;
 }
 
@@ -341,18 +342,25 @@ footer {
 
 .add-button {
     /* Remove the flex-grow property if you don't want the buttons to grow */
-    padding: 2px 8px; /* Adjust padding to fit the text */
+    padding: 2px 8px;
+    /* Adjust padding to fit the text */
     background-color: rgb(238, 238, 238);
-    border: none; /* Remove border if you don't need it */
-    cursor: pointer; /* Makes it clear the element is clickable */
-    white-space: nowrap; /* Prevent text inside the button from wrapping */
+    border: none;
+    /* Remove border if you don't need it */
+    cursor: pointer;
+    /* Makes it clear the element is clickable */
+    white-space: nowrap;
+    /* Prevent text inside the button from wrapping */
     /* You can remove min-width if you want the button to only be as wide as its content plus padding */
     /* min-width: 120px; */
-    margin: 2px; /* Provide some space around the buttons */
-    border-radius: 4px; /* If you want rounded corners */
+    margin: 2px;
+    /* Provide some space around the buttons */
+    border-radius: 4px;
+    /* If you want rounded corners */
     /* Add text alignment and other styles as needed */
     text-align: center;
-    font-size: 1rem; /* Adjust font size as needed */
+    font-size: 1rem;
+    /* Adjust font size as needed */
 }
 
 .add-button:hover {
