+++ src/views/EditProfileView.svelte
@@ -20,11 +20,6 @@
     let banner = "";
     let git_username = "";
     let git_proof = "";
-    let relays = [];
-
-    function handleRelayUpdate(event) {
-        relays = event.detail.relays;
-    }
 
     onMount(() => {
         if ($nostrManager) {
@@ -66,13 +61,10 @@
             git_username = profile.githubUsername;
             git_proof = profile.githubProof;
         }
-        relays = await fetchRelays();
-        console.log("existingRelays:", relays);
     }
 
     async function updateProfile() {
         if (!$nostrManager || !$nostrManager.write_mode) return;
-        $nostrManager.updateRelays(relays);
 
         const profileEvent = {
             kind: 0,
@@ -96,71 +88,6 @@
         } catch (error) {
             console.error("Error updating profile:", error);
         }
-        updateRelays();
-    }
-
-    function areSetsEqual(set1, set2) {
-        if (set1.size !== set2.size) return false;
-        for (let item of set1) {
-            if (!set2.has(item)) return false;
-        }
-        return true;
-    }
-
-    async function updateRelays() {
-        if (!$nostrManager || !$nostrManager.write_mode) return;
-
-        // Holen Sie die aktuellen Relays aus dem Cache
-        const existingRelays = await fetchRelays();
-        console.log("existingRelays:", existingRelays);
-
-        const existingRelaysSet = new Set(existingRelays);
-        const relaysSet = new Set(relays);
-        // Relay zum Array hinzufügen
-        const updatedRelays = relays;
-        console.log("updatedRelays:", updatedRelays);
-        // Überprüfen, ob das Relay bereits existiert
-        if (areSetsEqual(existingRelaysSet, relaysSet)) return;
-
-        // Event für die Aktualisierung der Relay-Liste erstellen
-        const relayEvent = createRelayEvent(updatedRelays);
-        console.log("relayEvent:", relayEvent);
-
-        // Event senden
-        try {
-            await $nostrManager.sendEvent(
-                relayEvent.kind,
-                relayEvent.content,
-                relayEvent.tags,
-            );
-            console.log("Relay added successfully");
-        } catch (error) {
-            console.error("Error adding relay:", error);
-        }
-    }
-
-    // Hilfsfunktion zum Erstellen eines Relay-Events
-    function createRelayEvent(relays) {
-        const content = ""; // Leerer Inhalt für Relay-Liste
-        const tags = relays.map((relay) => ["r", relay]);
-
-        return {
-            kind: 10002, // Kind für Relay-Liste
-            content,
-            tags,
-        };
-    }
-
-    // Hilfsfunktion zum Abrufen der aktuellen Relay-Liste aus dem Cache
-    async function fetchRelays() {
-        const relayEvents = await $nostrCache.getEventsByCriteria({
-            kinds: [10002],
-            authors: [$nostrManager.publicKey],
-        });
-
-        return relayEvents.flatMap((event) =>
-            event.tags.filter((tag) => tag[0] === "r").map((tag) => tag[1]),
-        );
     }
 
     let contentContainerClass = "combined-content-container";
@@ -210,7 +137,10 @@
                     style="top: -90px; position: relative"
                 >
                     <h2 class="base-h2 text-color-df">Edit Profile</h2>
-                    <div class="single-card-content text-color-df">
+                    <div
+                        class="single-card-content text-color-df"
+                        style="margin-bottom: 0;"
+                    >
                         <h5 class="base-h5 text-color-df">Name</h5>
                         <input
                             type="text"
@@ -270,10 +200,7 @@
                             style="width: 90%; margin: auto; margin-top: 30pt"
                         />
 
-                        <RelaySelectionWidget
-                            {relays}
-                            on:updateRelays={handleRelayUpdate}
-                        />
+                        <RelaySelectionWidget {profile_id} />
                     </div>
                 </div>
                 <div class="container mx-auto px-4 py-4 flex justify-end">
